requirements.txt
pandas
requests
python-dotenv
gspread_pandas
